Add files via upload
A classification prediction model is created using supervised machine learning where the model is
trained first on the training data set and then tested and validated on the test data set. Then
hyperparameter tuning is performed to find right combination of batch size and number of epochs.
Then optimization tuning is carried out to find the appropriate optimizer. Finally, the values are
predicted and training and validation loss is plotted.